feat(core variables): rename core warning to danger and added warning
Warning in B3 core related to danger in UIkit and there was no warning. This was confusing as shit. I renamed it to danger and added a warning color since I wanted to use it in the Project Assessment Tool and it's better to have it and not need it then need it and not have it I reckon.