fix(fonts): use CDN for open sans
We decided to do this to make it easier for both webpack & sprockets projects to use the fonts. Otherwise there's a lot of explaining and config you need to do for it to all work. If the project doesn't want a CDN, then they can add Open Sans in as part of their asset pipeline.